Spike showing restricted S3 access using cognito identity
This is to show how to restrict user access to only S3 objects that they have uploaded to prevent overwriting of other users' objects in S3

Keycloak is used as a developer authenticated identity in Cognito.

The identity is used to retrieve temporary credentials to perform S3 actions. These actions are restricted by an IAM policy that only allows the user to perform actions on S3 objects pre-fixed with the federated identity id

In addition to the code the following needs to be setup in AWS:
1. Cognito identity pool with Custom Authenticated provider name as: auth.tdr-integration.nationalarchives.gov.uk
2. IAM policy attached to the authenticated user IAM role to restrict access to S3 objects using: ${cognito-identity.amazonaws.com:sub}
3. AWS user with permission to access Cognito Client